Finished both outbound, and inbound communication. Pending multimedia provider. Eveything is NOT TESTED

+++ app/src/main/java/networker/RoomKnowledge.java
@@ -56,7 +56,7 @@ public class RoomKnowledge {
     public void incrementMessageReceived() {
         totalMessagesReceived.incrementAndGet();
     }
-    public void increaseContentSizeReceived(int byteCount) {
+    public void increaseContentSizeReceived(long byteCount) {
         totalContentSizeReceived.getAndAdd(byteCount);
     }
     public int getMessageReceivedCount() {
