+++ app/src/main/java/networker/messages/content/ContentProvider.java
@@ -3,5 +3,5 @@ package networker.messages.content;
 public interface ContentProvider<H, D> {
     H getHeader(); //TODO implement correctly in implementing classes
     D getData(); //TODO implement correctly in implementing classes
-    int getTotalSize(); //TODO implement correctly in implementing classes
+    long getTotalSize(); //TODO implement correctly in implementing classes
 }
